use nbt keys and put move back
