+++ src/core/java/com/enderio/core/CoreNBTKeys.java
@@ -11,6 +11,7 @@ public class CoreNBTKeys {
     public static final String ITEMS = "Items";
     public static final String FLUID = "Fluid";
     public static final String FLUIDS = "Fluids";
+    public static final String TANKS = "Tanks";
     public static final String ENERGY = "Energy";
     public static final String BLOCK_ENTITY_TAG = BlockItem.BLOCK_ENTITY_TAG;
